+++ src/Api/Providers/Pins.php
@@ -30,6 +30,7 @@ class Pins extends EntityProvider
         'create',
         'delete',
         'activity',
+        'analytics',
         'visualSimilar',
     ];
 
@@ -358,6 +359,16 @@ class Pins extends EntityProvider
         ];
 
         return $this->paginate($data, UrlBuilder::RESOURCE_EXPLORE_PINS, $limit);
+    }
+
+    /**
+     * Get pin analytics, like numbers of clicks, views and repins
+     * @param $pinId
+     * @return array|bool|Response
+     */
+    public function analytics($pinId)
+    {
+        return $this->get(['pin_id' => (string)$pinId], UrlBuilder::RESOURCE_PIN_ANALYTICS);
     }
     
     /**